@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.27)
 project(cpp17_concurrency_practice)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 
 add_executable(cpp17_concurrency_practice main.cpp
         chapter-1/1_hello_world.h
