@@ -6,8 +6,8 @@ namespace chapter_3 {
 #include <memory>
 #include <mutex>
 
-std::shared_ptr<int> resource_ptr;
-std::mutex resource_mutex;
+static std::shared_ptr<int> resource_ptr = nullptr;
+static std::mutex resource_mutex;
 
 /*!
  * @brief Так делать нельзя, большая вероятность
@@ -18,12 +18,18 @@ void undefined_behaviour_with_double_checked_locking() {
 	if (!resource_ptr) {
 		std::unique_lock<std::mutex> lock(resource_mutex);
 		if (!resource_ptr) {
-			resource_ptr.reset(new int(42));
+            /* Компилятор может сначала выделить память и назначить
+             * эту память переменной, а зетем уже записать туда 42 => возможна
+             * ситуация, когда resource_ptr уже не nullptr, но и не 42
+             * => если читать\писать без мьютекса - то это UB
+             */
+			resource_ptr = std::make_shared<int>(42);
 		}
 	}
 	++(*resource_ptr);
 }
 
+// Good code, example 1:
 std::once_flag resource_flag;
 void init_resource() {
 	resource_ptr.reset(new int(42));
@@ -38,6 +44,7 @@ void foo() {
 	++(*resource_ptr);
 }
 
+// Good code, example 2:
 class my_class;
 my_class& get_my_class_instance() {
 	// Начиная с C++11 данный код будет гарантированно
