@@ -10,7 +10,8 @@ namespace chapter_4 {
 
 /*!
  * @brief Потокобезопасный контейнер, который обеспечивает удобный
- * интерфейс.
+ * интерфейс без блокировок на ожидание мьютекса. Вместо блокировок
+ * здесь - сон.
  * wait_and_pop() и try_pop() - главные функции удобного интерфейса
  */
 template<typename T>
