@@ -1,14 +1,27 @@
 #ifndef INC_8_LATCHES_H
 #define INC_8_LATCHES_H
 
+#include <latch>
+#include <vector>
+#include <future>
+#include <thread>
+#include <iostream>
+
 namespace chapter_4 {
-/*!
- * std::latch is in std from C++20
- */
+
+using my_data = std::vector<unsigned int>;
+void do_more_stuff() { std::cout << "do_more_stuff()"; }
+my_data make_data(unsigned int i)
+{
+    return { 1 * i, 2 * i, 3 * i,
+             4 * i, 5 * i, 6 * i };
+}
+void process_data(my_data* data, unsigned int thread_count) { /* ... */ }
+
 void foo()
 {
-	unsigned const thread_count =...;
-	latch done(thread_count);
+	unsigned const thread_count = 8;
+	std::latch done(thread_count);
 	my_data data[thread_count];
 	std::vector<std::future<void>> threads;
 	for (unsigned i = 0; i < thread_count; ++i)
